# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Canadian Plans API",
    "version": "1.0.0",
    "description": "Generated from @canadian-plans/contracts Zod schemas. Do not edit by hand — run `pnpm --filter @canadian-plans/contracts build`."
  },
  "servers": [{ "url": "/", "description": "Backend origin; all routes are under /api/v1." }],
  "components": {
    "securitySchemes": {
      "staffSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase Auth access token; verified and re-authorised on every request."
      },
      "websiteCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-site service credential (`cplsk_…`); resolves to one workspace and its scopes."
      },
      "machineSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "Provider HMAC signature verified through the server-only machine registry."
      }
    }
  },
  "paths": {
...
    "/api/v1/website/leads": {
      "post": {
        "operationId": "createLead",
        "summary": "Create a lead draft with attribution.",
        "tags": ["website"],
        "security": [{ "websiteCredential": [] }],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lead": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "workspaceId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "status": { "type": "string", "enum": ["incomplete", "submitted"] },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        }
                      },
                      "required": ["id", "workspaceId", "status", "updatedAt"],
                      "additionalProperties": false
                    },
                    "draftGrant": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "minLength": 1 },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        }
                      },
                      "required": ["token", "expiresAt"],
                      "additionalProperties": false
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["lead", "draftGrant", "requestId"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error envelope `{ error: { code, message, requestId, details? } }`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "invalid_request",
                            "missing_session",
                            "invalid_session",
                            "workspace_not_found",
                            "membership_missing",
                            "membership_pending",
                            "membership_revoked",
                            "permission_denied",
                            "mfa_required",
                            "membership_not_found",
                            "membership_conflict",
                            "internal_error",
                            "missing_credential",
                            "invalid_credential",
                            "credential_revoked",
                            "credential_not_found",
                            "scope_denied",
                            "caller_forbidden",
                            "rate_limited",
                            "payload_too_large",
                            "machine_unknown",
                            "machine_signature_invalid",
                            "machine_account_mismatch",
                            "machine_workspace_mismatch",
                            "validation_error",
                            "not_found",
                            "conflict",
                            "idempotency_conflict",
                            "draft_not_found",
                            "draft_expired",
                            "offer_unavailable",
                            "quote_not_found",
                            "quote_expired",
                            "quote_withdrawn",
                            "order_not_found",
                            "version_conflict",
                            "terms_version_unsupported",
                            "upload_not_found",
                            "upload_invalid",
                            "unsupported_media_type",
                            "file_not_found",
                            "download_denied",
                            "commission_not_found",
                            "commission_state_invalid",
                            "invoice_not_found",
                            "invoice_conflict",
                            "export_denied",
                            "tracking_challenge_invalid",
                            "tracking_expired",
                            "tracking_attempts_exceeded",
                            "webhook_rejected"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "details": {
                          "type": "object",
                          "propertyNames": { "type": "string" },
                          "additionalProperties": {}
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attribution": {
                    "type": "object",
                    "propertyNames": { "type": "string" },
                    "additionalProperties": {}
                  },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "fullName": { "type": "string", "minLength": 1, "maxLength": 160 },
                      "email": {
                        "type": "string",
                        "maxLength": 254,
                        "format": "email",
                        "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "phone": { "type": "string", "maxLength": 32 },
                      "countryCode": { "type": "string", "pattern": "^[A-Z]{2}$" }
                    }
                  },
                  "form": {
                    "type": "object",
                    "properties": {
                      "schemaVersion": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      "payload": {
                        "type": "object",
                        "propertyNames": { "type": "string" },
                        "additionalProperties": {}
                      }
                    },
                    "required": ["schemaVersion", "payload"]
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "consentVersion": { "type": "string", "minLength": 1, "maxLength": 64 }
                },
                "required": ["consentVersion"]
              }
            }
          }
        }
      }
    },
    "/api/v1/website/leads/{leadId}": {
      "patch": {
        "operationId": "updateLead",
        "summary": "Resume and update a draft.",
        "tags": ["website"],
        "security": [{ "websiteCredential": [] }],
        "parameters": [
          { "name": "leadId", "in": "path", "required": true, "schema": { "type": "string" } },
          {
            "name": "X-Draft-Grant",
            "in": "header",
            "required": true,
            "description": "Scoped grant token that authorises resuming this draft.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lead": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "workspaceId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "status": { "type": "string", "enum": ["incomplete", "submitted"] },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        }
                      },
                      "required": ["id", "workspaceId", "status", "updatedAt"],
                      "additionalProperties": false
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["lead", "requestId"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error envelope `{ error: { code, message, requestId, details? } }`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "invalid_request",
                            "missing_session",
                            "invalid_session",
                            "workspace_not_found",
                            "membership_missing",
                            "membership_pending",
                            "membership_revoked",
                            "permission_denied",
                            "mfa_required",
                            "membership_not_found",
                            "membership_conflict",
                            "internal_error",
                            "missing_credential",
                            "invalid_credential",
                            "credential_revoked",
                            "credential_not_found",
                            "scope_denied",
                            "caller_forbidden",
                            "rate_limited",
                            "payload_too_large",
                            "machine_unknown",
                            "machine_signature_invalid",
                            "machine_account_mismatch",
                            "machine_workspace_mismatch",
                            "validation_error",
                            "not_found",
                            "conflict",
                            "idempotency_conflict",
                            "draft_not_found",
                            "draft_expired",
                            "offer_unavailable",
                            "quote_not_found",
                            "quote_expired",
                            "quote_withdrawn",
                            "order_not_found",
                            "version_conflict",
                            "terms_version_unsupported",
                            "upload_not_found",
                            "upload_invalid",
                            "unsupported_media_type",
                            "file_not_found",
                            "download_denied",
                            "commission_not_found",
                            "commission_state_invalid",
                            "invoice_not_found",
                            "invoice_conflict",
                            "export_denied",
                            "tracking_challenge_invalid",
                            "tracking_expired",
                            "tracking_attempts_exceeded",
                            "webhook_rejected"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "details": {
                          "type": "object",
                          "propertyNames": { "type": "string" },
                          "additionalProperties": {}
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contact": {
                    "type": "object",
                    "properties": {
                      "fullName": { "type": "string", "minLength": 1, "maxLength": 160 },
                      "email": {
                        "type": "string",
                        "maxLength": 254,
                        "format": "email",
                        "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "phone": { "type": "string", "maxLength": 32 },
                      "countryCode": { "type": "string", "pattern": "^[A-Z]{2}$" }
                    }
                  },
                  "form": {
                    "type": "object",
                    "properties": {
                      "schemaVersion": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      "payload": {
                        "type": "object",
                        "propertyNames": { "type": "string" },
                        "additionalProperties": {}
                      }
                    },
                    "required": ["schemaVersion", "payload"]
                  },
                  "attribution": {
                    "type": "object",
                    "propertyNames": { "type": "string" },
                    "additionalProperties": {}
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "consentVersion": { "type": "string", "minLength": 1, "maxLength": 64 }
                }
              }
            }
          }
        }
      }
    },
...
  }
}